Feature/5 complete presentation layer

+++ data/src/main/java/com/example/data/daos/NoteDao.kt
@@ -26,5 +26,5 @@ interface NoteDao {
     suspend fun getNotes(): List<NoteEntity>
 
     @Query("SELECT * FROM notes WHERE id = :id")
-    suspend fun getNoteById(id: Long): NoteEntity?
+    suspend fun getNoteById(id: Int): NoteEntity?
 }
